Add debug mode to test Neon Racer without server
