+++ src/main/java/neonracer/NeonRacerMain.java
@@ -17,6 +17,7 @@ public class NeonRacerMain {
 
         // Create a game context
         GameContext gameContext = GameContextFactory.createDefault();
+        //GameContext gameContext = GameContextFactory.createForDebug();
         gameContext.initialize();
 
         // Create the master renderer and start the rendering
