+++ src/main/java/neonracer/core/GameContextFactory.java
@@ -1,6 +1,7 @@
 package neonracer.core;
 
-import neonracer.client.Client;
+import neonracer.client.DebugClient;
+import neonracer.client.NetworkClient;
 import neonracer.gui.input.KeyboardState;
 import neonracer.gui.input.MouseState;
 import neonracer.phys.PhysicsEngine;
@@ -22,8 +23,22 @@ public class GameContextFactory {
         Timer timer = new Timer(60);
         MouseState mouseState = new MouseState();
         KeyboardState keyboardState = new KeyboardState();
-        Client client = new Client();
-        return new GameContext(gameWindow, textureProvider, shaderProvider, dataManager, gameState, physicsEngine, timer, mouseState, keyboardState, client);
+        NetworkClient client = new NetworkClient();
+        return new GameContext(gameWindow, textureProvider, shaderProvider, dataManager, gameState, physicsEngine, timer, mouseState, keyboardState, client, false);
+    }
+
+    public static GameContext createForDebug() {
+        GameWindow gameWindow = new GameWindow(800, 600, BuildInfo.getGameTitle() + " - Debug");
+        TextureProvider textureProvider = new TextureProvider();
+        ShaderProvider shaderProvider = new ShaderProvider();
+        DataManager dataManager = new DataManager();
+        GameState gameState = new GameState();
+        PhysicsEngine physicsEngine = new PhysicsEngine();
+        Timer timer = new Timer(60);
+        MouseState mouseState = new MouseState();
+        KeyboardState keyboardState = new KeyboardState();
+        DebugClient client = new DebugClient();
+        return new GameContext(gameWindow, textureProvider, shaderProvider, dataManager, gameState, physicsEngine, timer, mouseState, keyboardState, client, true);
     }
 
     public static GameContext createForDesigner() {
@@ -35,7 +50,7 @@ public class GameContextFactory {
         PhysicsEngine physicsEngine = new PhysicsEngine();
         MouseState mouseState = new MouseState();
         KeyboardState keyboardState = new KeyboardState();
-        return new GameContext(gameWindow, textureProvider, shaderProvider, dataManager, gameState, physicsEngine, null, mouseState, keyboardState, null);
+        return new GameContext(gameWindow, textureProvider, shaderProvider, dataManager, gameState, physicsEngine, null, mouseState, keyboardState, null, false);
     }
 
 }
